[cpp][jinja][level] Convert the "StageSelectComponent" to auto-gen "GameStateLevelComponent" jinja files.
This is mainly to experiment with Jinja/AutoGen, not very practical. It does generalize the component class more - so that designers can define which game state class to have the component push/pop - but this could maybe be achieved by other means (c++ class template maybe, or AzCore's RTTI to choose a type by type id at runtime.